requirements.txt
dvc
tqdm
Whoosh
pandas
praw
python-dotenv
flask
flask-RESTful
flask-cors
markdown
beautifulsoup4